Move timesorter to its own package and remove unused functions
